refactor : domain관련 폴더 분리
