+++ src/components/Modal/Modal.style.ts
