+++ src/domains/ShoppingProducts/apis/mocks/mockProductResponse.ts
