+++ src/domains/ShoppingProducts/apis/product.ts
