+++ src/domains/ShoppingProducts/apis/types/product.ts
