+++ src/domains/ShoppingProducts/apis/utils/apiUtils.ts
