+++ src/domains/ShoppingProducts/components/QuantitySelector/QuantitySelector.tsx
@@ -1,5 +1,5 @@
-import { updateCartItem } from "../../api/cartItem";
-import { IconButton } from "../IconButton/IconButton";
+import { updateCartItem } from "../../apis/cartItem";
+import { IconButton } from "../../../../components/IconButton/IconButton";
 import { QuantitySelectorLayout } from "./QuantitySelector.style";
 
 interface QuantitySelectorProps {
