+++ src/mocks/node.ts
@@ -1,4 +1,4 @@
 import { setupServer } from "msw/node";
-import { handlers } from "./apiHandlers";
+import { handlers } from "../domains/ShoppingProducts/apis/mocks/apiHandlers";
 
 export const server = setupServer(...handlers);
